bot messages with middleware
