+++ Valentina_Aktanko/src/middlewares/logger.js
@@ -1,0 +1,13 @@
+export function logger(store) {
+    return function wrapDispatch(next) {
+        return function dispatchAndLog(action) {
+            console.log('Action' , action);
+            console.log('prevState', store.getState().chats.toJS());
+
+            const result = next(action);
+            console.log('nextState', store.getState().chats.toJS());
+
+            return result;
+        }
+    }
+}
